+++ xash3d_engine/engine/ref/soft/r_math.c
@@ -125,7 +125,7 @@ void Matrix4x4_CreateRotate(matrix4x4 out, float angle, float x, float y, float 
 
 	len = x * x + y * y + z * z;
 	if ( len != 0.0f )
-		len = 1.0f / sqrt(len);
+		len = 1.0f / sqrtf(len);
 	x *= len;
 	y *= len;
 	z *= len;
